refactor(kml): split point_extractor.py into one file per class
- pixel_point.py: PixelPoint dataclass
- kml_point.py: KmlPoint dataclass with from_placemark factory
- kml.py: Kml parser class
- geo_config.py: GeoConfig and Geotransform type alias
- geo_point_registry.py: GeoPointRegistry (renamed from PointExtractor)
