+++ poc_homography/kml/__init__.py
@@ -1,12 +1,9 @@
 """KML utilities for georeferenced point extraction and export."""
 
-from poc_homography.kml.point_extractor import (
-    GeoConfig,
-    Geotransform,
-    Kml,
-    KmlPoint,
-    PixelPoint,
-    PointExtractor,
-)
+from poc_homography.kml.geo_config import GeoConfig, Geotransform
+from poc_homography.kml.geo_point_registry import GeoPointRegistry
+from poc_homography.kml.kml import Kml
+from poc_homography.kml.kml_point import KmlPoint
+from poc_homography.kml.pixel_point import PixelPoint
 
-__all__ = ["GeoConfig", "Geotransform", "Kml", "KmlPoint", "PixelPoint", "PointExtractor"]
+__all__ = ["GeoConfig", "Geotransform", "GeoPointRegistry", "Kml", "KmlPoint", "PixelPoint"]
